♻️ refactor: Profile 리팩토링
