+++ src/components/profile/CoverImage/CoverImage.tsx
