+++ src/components/profile/CoverImage/index.ts
@@ -1,0 +1,3 @@
+import CoverImage from "./CoverImage";
+
+export default CoverImage;
