+++ src/components/profile/CoverImage/style.ts
