+++ src/components/profile/CoverImageBox/CoverImageBox.tsx
