+++ src/components/profile/CoverImageBox/index.ts
@@ -1,0 +1,3 @@
+import CoverImageBox from "./CoverImageBox";
+
+export default CoverImageBox;
