+++ src/components/profile/CoverImageBox/style.ts
