+++ src/components/profile/FollowIcon/CreateFollowIcon.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { ReactComponent as CreateFollowSVG } from "@assets/icons/icon_create_follow.svg";
-import { Id } from "../ProfileImageBox/index";
+import { Id } from "../ProfileImageBox/ProfileImageBox";
 
 const CreateFollowIcon: FC<{ id: Id; onCreate }> = ({ id, onCreate }) => {
   return (
