+++ src/components/profile/FollowIcon/index.ts
@@ -1,2 +1,3 @@
 import FollowIcon from "./FollowIcon";
+
 export default FollowIcon;
