+++ src/components/profile/ProfileImageBox/index.ts
@@ -1,0 +1,3 @@
+import ProfileImageBox from "./ProfileImageBox";
+
+export default ProfileImageBox;
