+++ src/components/profile/Tab/index.ts
@@ -1,0 +1,3 @@
+import Tab from "./Tab";
+
+export default Tab;
