+++ src/components/profile/index.ts
@@ -1,19 +1,19 @@
 import CoverImage from "./CoverImage";
 import CoverImageBox from "./CoverImageBox";
-import Modal from "./Modal";
-import Tab from "./Tab";
-import ProfileImageBox from "./ProfileImageBox";
 import EditFullName from "./EditFullName";
 import EditPassword from "./EditPassword";
 import FollowIcon from "./FollowIcon";
+import Modal from "./Modal";
+import ProfileImageBox from "./ProfileImageBox";
+import Tab from "./Tab";
 
 export {
   CoverImage,
   CoverImageBox,
-  Modal,
-  ProfileImageBox,
-  Tab,
   EditFullName,
   EditPassword,
-  FollowIcon
+  FollowIcon,
+  Modal,
+  ProfileImageBox,
+  Tab
 };
