+++ src/pages/profile/index.ts
@@ -1,0 +1,3 @@
+import profile from "./profile";
+
+export default profile;
